clean code (Success)

+++ src/main/java/com/example/BackendApp/service/UserService.java
@@ -4,7 +4,6 @@ import com.example.BackendApp.entity.UserEntity;
 import com.example.BackendApp.model.*;
 import com.example.BackendApp.repository.UserRepo;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -26,34 +25,28 @@ public class UserService {
         return userRepo.findAll();
     }
 
-    public ResponseEntity<String> createUser(UserRequest userRequest) {
+    public ResponseEntity<String> createUser(RegistrationRequest registrationRequest) {
         try {
-            int nameLength = userRequest.getUserName().length();
-            int emailLength = userRequest.getEmail().length();
+            int nameLength = registrationRequest.getUserName().length();
+            int emailLength = registrationRequest.getEmail().length();
 
             for (int i = 0; i < nameLength; i++) {
-                if (userRequest.getUserName().charAt(i) == ' ')
+                if (registrationRequest.getUserName().charAt(i) == ' ')
                     return ResponseEntity.badRequest().body("There should be no spaces in the name");
             }
             for (int i = 0; i < emailLength; i++) {
-                if (userRequest.getEmail().charAt(i) == ' ') {
+                if (registrationRequest.getEmail().charAt(i) == ' ') {
                     return ResponseEntity.badRequest().body("There should be no spaces in the email");
                 }
             }
-            if (userRepo.findByEmail(userRequest.getEmail()) != null) {
+            if (userRepo.findByEmail(registrationRequest.getEmail()) != null) {
                 return ResponseEntity.badRequest().body("A user with this email already exists!");
             }
-            if (!Objects.equals(userRequest.getConfirmPass(), userRequest.getUserPass())) {
+            if (!Objects.equals(registrationRequest.getConfirmPass(), registrationRequest.getUserPass())) {
                 return ResponseEntity.badRequest().body("The repeated password does not match the current password");
             }
-            UserEntity user = new UserEntity();
-            LocalDateTime dateOfRegister = LocalDateTime.now();
-            user.setDateOfRegister(dateOfRegister);
-            user.setEmail(userRequest.getEmail());
-            user.setUserName(userRequest.getUserName());
-            user.setUserPass(passwordEncoder.encode(userRequest.getUserPass()));
-            user.setRole(Role.valueOf("USER"));
-            user.setStatus(Status.valueOf("ACTIVE"));
+            UserEntity user = new UserEntity(registrationRequest.getEmail(), registrationRequest.getUserName(), passwordEncoder.encode(registrationRequest.getUserPass()),
+                    LocalDateTime.now(), Role.USER, Status.ACTIVE);
             userRepo.save(user);
             return new ResponseEntity<String>("User is created", HttpStatus.CREATED);
         } catch (Exception e) {
@@ -61,37 +54,33 @@ public class UserService {
         }
     }
 
-    public ResponseEntity<?> updateUser(Long id, UserRequest userRequest) {
+    public ResponseEntity<?> updateUser(Long id, RegistrationRequest registrationRequest) {
         return userRepo.findById(id)
                 .map(user1 -> {
-                    LocalDateTime dateAfterUpdate = LocalDateTime.now();
-                    user1.setUserName(userRequest.getUserName());
-                    user1.setEmail(userRequest.getEmail());
-                    user1.setDateOfRegister(dateAfterUpdate);
-                    user1.setUserPass(passwordEncoder.encode(userRequest.getUserPass()));
-                    user1.setRole(Role.valueOf("USER"));
-                    user1.setStatus(Status.valueOf("ACTIVE"));
+                    user1.setUserName(registrationRequest.getUserName());
+                    user1.setEmail(registrationRequest.getEmail());
+                    user1.setDateOfRegister(LocalDateTime.now());
+                    user1.setUserPass(passwordEncoder.encode(registrationRequest.getUserPass()));
+                    user1.setRole(Role.USER);
+                    user1.setStatus(Status.ACTIVE);
                     userRepo.save(user1);
                     return ResponseEntity.ok("A user with such an ID " + id + " updated");
                 }).orElse(new ResponseEntity<String>("A user with such an ID " + id + " not found", HttpStatus.NOT_FOUND));
     }
 
     public ResponseEntity<?> getUserId(Long id) {
-        try {
-            UserEntity user = userRepo.findById(id).get();
-            return ResponseEntity.ok(UserModel.toUser(user));
-        } catch (Exception e) {
+        if (userRepo.findById(id).isPresent()) {
+            return ResponseEntity.ok(UserModel.toUser(userRepo.findById(id).get()));
+        } else {
             return new ResponseEntity<String>("A user with such an ID " + id + " not found", HttpStatus.NOT_FOUND);
         }
     }
 
     public ResponseEntity<String> deleteUser(Long id) {
-        try {
+        if (userRepo.existsById(id)) {
             userRepo.deleteById(id);
             return ResponseEntity.ok("User is deleted");
-        } catch (Exception e) {
-            return new ResponseEntity<String>("A user with such an ID " + id + " not found", HttpStatus.NOT_FOUND);
         }
+        else return new ResponseEntity<String>("There is no such user", HttpStatus.NOT_FOUND);
     }
-
 }
